InitialNavigationDispatcher: Clarify and improve `doInitialNavigation`.
When the active account isn't logged in and there's just one account
in total, we can reasonably pre-fill that account's realm in the
realm-input screen. It seems like this has been the expected
behavior [1], but I just tested it now, and it hasn't been
happening.

Also, for readability, separate the `accounts.length === 1` check
into its own `else if`.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/decouple.20nav.20from.20redux.20%28.23M3804%29/near/1042602