android: Request legacy external storage.
This opts out of Android 10's new "scoped storage" model.  Docs:
  https://developer.android.com/training/data-storage/use-cases#opt-out-scoped-storage

The new model otherwise takes effect on Android 10+ devices when
targeting Android 10+.  We're about to start targeting Android 10
(#3665), because Google Play will require us to do so starting on
2020-11-02, which is soon.

The opt-out will stop having any effect on Android 11+, when we
eventually target Android 11+.  So effectively it gives us an
extension of one more cycle on adapting to the new model.

The reason we're opting out is that at least one of our dependencies,
the react-native-image-picker library which we use for taking photos
and also picking existing photos to share on Zulip, hasn't yet
adapted to work properly with scoped storage; if you try doing so in
a Zulip that targets Android 10 but doesn't have this change, you
just get an error.  I've filed #4283 with details of the issue and
what needs to be done to properly fix it.