narrow tests: Make tests well-typed.
Mostly this involves using real Message objects instead of objects
with a handful of the same properties, and passing the full list of
arguments to some function calls where the tests were getting away
without the last one because the control path we're exercising
happens not to use it.

We also discard some test fragments that just check behavior on
ill-typed data that doesn't particularly correspond to any data we
believe our actual code produces, like `[{}, {}]` to the isFooNarrow
functions or `undefined` to `isSameNarrow`.

There's a lot more that could be done to simplify these tests and
further improve them, but this gives us a basis for any refactoring.