narrowsReducer tests: Fix a tangled-up test case.
The data in this test case has never made much sense since 7a5da9587,
shortly after the test was added -- testing a PM with a sender not in
the display_recipient, which never happens.  With e09648c68 it grew
another kind of testing mistake which helped mask that problem: the
`caughtUp` state was false for the narrows we didn't expect the
message to get added to, so we didn't actually test whether the logic
believed the message belonged to them.

Then a bug in eg.pmMessage made the message differently malformed,
after 4077e8880 had this test start using that function; the message
would match all PM narrows, and the test kept passing only because of
that `caughtUp` masking.

We're about to fix that bug in eg.pmMessage.  That will cause this
test to fail (with the message matching no conversation narrows)
unless adjusted.  Just fix the whole test to make more sense.