recipient: Factor out helper pmKeyRecipientsFromIds.
Does much the same job as pmKeyRecipientsFromMessage, but starting
from a list of user IDs.  We have this logic already to handle
notifications for group PMs, which come with such a list encoded as
comma-separated strings.  We'll want similar logic for group-PM narrow
URLs, which do the same thing; so factor it out to be shared.