outbox [nfc]: Add a TODO to switch to numeric IDs when sending.
These are only supported beginning with v2.0 of the server, so we can
only do this either conditioned on the server version, or after older
versions are no longer supported.

This isn't something where the old way is deprecated and we're eager
to remove it from a future server version -- to the contrary, we
intend to support stream names and user emails here forever, as part
of keeping easy things easy when writing bots.  So there isn't a lot
to be gained by eagerly switching; the main thing is it'd mitigate a
race where a user we're sending to happens to have changed their email
address right as we're trying to send.