narrow [nfc]: Use keyFromNarrow rather than direct JSON.stringify.
Done mostly automatically, as follows:

  $ perl -i -0pe '
      if (s<\bJSON\.stringify(\(.*(?i)narrow)>
           <keyFromNarrow\1>g) {
        s<^import .*?\K,?\s*(\} from \S*/narrow.;)>
         <, keyFromNarrow \1>ms
        || s<^import .* from \S*;\K$>
            <\nimport { keyFromNarrow } from "../utils/narrow";>ms;
      }
    ' src/**/*.js
  $ git restore src/api/
  $ tools/fmt

The Perl snippet rewrites `JSON.stringify` calls that appear to be on
anything the code calls a "narrow", and then adds an import as needed.
We leave out a couple of matches in src/api/ which really aren't
about keys in our data structures but about communicating with the
server, so that `JSON.stringify` is the right name for what we're
doing.

Beyond the above commands, we make a couple of manual fixups just in
narrow.js itself.

Searching the codebase for uses of `JSON.stringify`, all of the
remaining ones appear to be about things other than narrows.