narrow: Encode our own structured keys instead of JSON.
The main motivation for this change is that it will help us switch to
a more-structured type for narrows in our code, rather than the
existing arrays of operator/operand pairs -- we'll no longer be
coupling our data structures to the details of those values.

One way this change could affect the code's behavior, in principle, is
that JSON.stringify emits an object's properties in enumeration order
-- which means that narrows we consider equal may not have gotten
serialized as equal:

  > JSON.stringify([{"operator": "stream", "operand": "general"}])
  '[{"operator":"stream","operand":"general"}]'

  > JSON.stringify([{"operand": "general", "operator": "stream"}])
  '[{"operand":"general","operator":"stream"}]'

which could result in looking for the same narrow at different keys.
If we have any cases where this can currently happen, then these
explicitly-structured keys fix that.