avatar: Handle neglected cases where server sends a Gravatar URL.
This fixes a bug stemming from the assumption that, if `avatar_url`
from the server is a string, it won't be a Gravatar URL. There are
in fact two cases where we'll get a string that represents a
Gravatar URL:

  - If we don't announce `client_gravatar`. Although we plan to fix
    it soon, we don't currently announce it in `getMessages`.

  - If the server doesn't have EMAIL_ADDRESS_VISIBILITY_EVERYONE
    set. When that's set, real email addresses don't get sent to the
    client. But real email addresses are still supposed to determine
    Gravatar URLs, since I might have told Gravatar what my email
    address is, and what image to show for me. The server can't
    expect the client to compute the Gravatar URL for the real email
    address, since the server isn't sending real email addresses. So
    they just give the client the Gravatar URL for the real email
    address.

Reading past the first conditional in `getAvatarUrl`, we appear to
have assumed that the `avatar_url` string would correspond to a
relative or absolute URL for an uploaded avatar. This code was
mostly harmless to a Gravatar URL string that had snuck through for
one of the above reasons. Gravatar URLs are absolute URLs, so
`realm` in `new URL(avatarUrl, realm)` would have been ignored. If
`size` was greater than 100, it would have been run through a regex,
but that's a no-op because Gravatar URLs don't have `.png` in them.
So, no runtime errors.

I say it was "mostly harmless" because there is some behavior we
didn't intend: a Gravatar URL string sneaking through in these cases
wouldn't ever be set to the desired or the Zulip-default size (80).
The way to set a Gravatar URL to a certain size is to pass a query
param `s`, like `s=100`. And that just wasn't being done in these
cases because the code wasn't written to be aware that it might be
working with a Gravatar URL.

So, fix that by recognizing when we're dealing with a Gravatar URL
the server gave us and sizing it appropriately.

With these Gravatar URLs being sized now -- with numbers chosen
correctly in the previous commit -- we can say #4305 is fixed. More
discussion of this is at at that issue and at
https://github.com/zulip/zulip-mobile/pull/4230#discussion_r523383345.

Fixes: #4305