narrow [nfc]: Rename `getNarrowFromMessage` to `getNarrowForReply`.
A message can easily exist in multiple narrows (e.g., a stream
narrow, a topic narrow, and the starred narrow), so
`getNarrowFromMessage` was a misleading name.

Looking at its implementation and its one call site, this function
does some important work to answer the reasonable question, "Which
narrow should my reply to this message go in?". So, name it that
way.