android: Follow RN's removal of `pickFirst` arguments.
Part of the RN v0.62 -> v0.63 changes to the template app [1],
corresponding to facebook/react-native@4bb0b4f20.

These lines (which we added in cb56f1646, for
facebook/react-native@2fd50882b) are reportedly "no longer necessary
with the new Flipper release".

It would have been nice if they specified which Flipper release. At
that commit, the Flipper version in use is 0.30.2, so it probably
would have been safe to make this change for quite a while -- but
might as well do it now, after having just taken a big leap in
Flipper versions well beyond that.

[1] https://react-native-community.github.io/upgrade-helper/?from=0.62.2&to=0.63.4
[2] https://github.com/facebook/react-native/blob/4bb0b4f20/template/android/gradle.properties#L28