users [nfc]: Rename some usersBy* -> allUsersBy* where that's already true.
This doesn't actually change what data we use in any of these
codepaths -- just the names.

Ultimately we probably want to be have just one map, namely
"all users, by ID", and drop the "all" from its name because we'll
stop having to think about the distinction.  But until we're using
"all users" maps everywhere, the names will help us track where we
are vs. where we aren't yet.

A few of these were recently changed to be "all users"; most of
them have been for a long time.