narrowsReducer: Fix a perf regression in EVENT_NEW_MESSAGE handling.
Thanks to Greg for the analysis, which I'm summarizing here, and the
suggested solution. For more background, see discussion [1].

When we started using `Immutable.Map` for `state.narrows` in a
recent commit, there was a performance regression on handling
EVENT_NEW_MESSAGE when `state.narrows` has very many keys (we tested
with 1000). We left the regression in, until now, so that commit
could remain a direct, easy-to-follow translation from plain
object-as-map to `Immutable.Map`.

Although using `Immutable.Map` the way we did had the advantage
(over an object-as-map, or over an ES6 `Map`) of not having to
allocate a whole new copy of `state.narrows` just to change one
thing, it still meant *iterating* through all the narrows to check
whether we needed to update each one. So the work being done was
still linear with the number of narrows, even supposing (which is
very reasonable) that the constant factor for read-only iterations
was much smaller than for iterations that had to allocate.

Greg then tried the approach in this commit, which isn't linear in
the number of narrows, and he found that it was significantly faster
for 1000 narrows. The key fact is that you can look at a message and
tell what narrows it belongs in, and that's always a very low number
(like five) [2].

It also makes an improvement that's orthogonal to `Immutable.Map`
vs. objects-as-map vs. `Map`: on an iteration, we should skip the
`messageDoesNotExist` computation (which goes and checks all items
in the array) if `isCaughtUp` has already indicated we're not
interested in that narrow [3].

See his observations for 1000 narrows [4] and for (the much more
usual) 30 narrows [5], where the approach in this commit still shows
up as faster than the object-as-map approach before this series.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60Immutable.2EMap.60.20performance.20for.20.60state.2Enarrows.60/near/1068889

[2] As long as you exclude search narrows, which there's a pretty
    much unbounded number of -- but we've never bothered
    live-updating them with message events anyway.

[3] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60Immutable.2EMap.60.20performance.20for.20.60state.2Enarrows.60/near/1067739

[4] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60Immutable.2EMap.60.20performance.20for.20.60state.2Enarrows.60/near/1068895

[5] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60Immutable.2EMap.60.20performance.20for.20.60state.2Enarrows.60/near/1068917