narrow tests: Add round-trip tests for keyFromNarrow and parseNarrow.
We have a couple of parseNarrow tests that start from specific
strings, but the real spec for this function is that it's the inverse
of keyFromNarrow.  Use that to make some much more thorough tests.

That's also basically the spec for keyFromNarrow, too.  (That, and
that it should be deterministic.)