tests: Stop constructing narrows as literals.
We're going to change the internal representation of narrows soon.
The internal representation isn't really any test's business in the
first place, and it especially was never the business of any of these
tests that aren't of the `narrow` module itself.  Instead, they should
use the proper constructors and constants from the `narrow` module.