user types: Switch to UserId in most React props and state.
All these changes are pretty boring.

Many of them are independent of each other, but others aren't because
one component passes a user ID to another, either as a child or by
navigation.

This change is limited to the places where a user ID, or list of them,
appears in order to specify particular users that this component, this
time, should do its work with respect to.  Upcoming commits will take
care of places where we pass general information that happens to
contain user IDs.