flipper ios: Follow RN to expose pod-conditioning by build configuration.
Part of the RN v0.62 -> v0.63 changes to the Flipper-pods logic,
transposed to the Podfile, corresponding to
facebook/react-native@c72ecdb90. See the last few commits for why
the upstream commit makes these changes somewhere other than the
template-app Podfile (and why the changes don't appear in the
upgrade guide [1]).

[1] https://react-native-community.github.io/upgrade-helper/?from=0.62.2&to=0.63.4