isInternalLink [nfc]: Make another simplification step.
If the URL starts with `/`, then it cannot match the branch of the
"#narrow" regexp that requires it to start with `#`.

If the URL starts with `http` or `WwW.` (... or for that matter `www!`
or `wwwz` or any of the other strings that match the regexp /www./i),
then it cannot match the "#narrow" regexp.  So we can go straight to
testing it against the "#narrow" regexp, and we'll get the same result
of `false` as the nonsense extra condition was giving us.