isInternalLink: Stop accepting URLs with just one extra slash.
Because the realm URL we use comes from the serialization of a
URL object, its path cannot be empty -- it can look like
`https://chat.example.org/`, or perhaps `https://chat.example.org///`,
but never `https://chat.example.org`.

As a result there's no need to make room for a potential leading slash
in what we get after stripping it off as a prefix.

This is a behavior change in that previously if the realm URL was
`https://chat.example/`, then we'd accept `https://chat.example//#narrow`,
and now we don't.  We already didn't accept a URL with more than one
extra slash, like `https://chat.example///#narrow`.  I'm not aware of
any reason to think that links with even one extra slash like this are
found in the wild; the old behavior was just a quirk of the old terse
and incompletely-thought-out logic.