store: Remove replace/revive handlers for Immutable things we don't use.
This has been observed to save >20% of the time it takes to persist
our Redux storage to disk [1].

It's also convenient for the mechanics of inlining
remotedev-serialize; e.g., we won't have to spell out `refer` at
all, since it's not used after this commit.

Originally opened as zulip/remotedev-serialize#2, before it became
clear that we'd like to inline remotedev-serialize.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/user_avatar_url_field_optional/near/1082087