example data: Add a batteries-included version of the Redux state.
As we start having reducers take the global state as an additional
argument, we're going to more frequently have a need for an example
global state that includes common data like the self user.  It's
already perhaps past time to have such a thing provided centrally in
exampleData.  Add one now.