ComposeBox: Stop fetching topics on every focus of topic input.
Before this, we were fetching the list of topics to populate
TopicAutocomplete more frequently than necessary; in particular, we
were doing so every time the topic input in `ComposeBox` was
focused.

The following should be sufficient to ensure we're up-to-date with
the complete list of topics at all times that we need to be:

- When we first expect to see the list, fetch all topics for the
  stream.

- Afterwards, update the list when a new message arrives, if it
  introduces a new topic.

The latter is already taken care of (see handling of
EVENT_NEW_MESSAGE in topicsReducer).

So, do the former -- and do it in `TopicAutocomplete`, which is much
closer to where we consume the data. (ComposeBox is getting pretty
crowded already.)

See discussion [1].

[1] https://github.com/zulip/zulip-mobile/pull/4351#issuecomment-749099486

Co-authored-by: Chris Bobbe <cbobbe@zulip.com>
Fixes: #2750