webview: Use native scroll deceleration on iOS
The default WebView `decelerationRate` is 'fast', which slows down
momentum scrolling through messages significantly. Change it to
'normal' which is the native rate on iOS and makes it behave like
other apps, allowing for much faster scrolling.

Test Plan: confirmed that this has the expected scroll momentum on an
iPhone 12.

Fixes: #3557