replaceRevive [nfc]: Use `origValue` instead of `value` in more places.
One of the `invariant`s just above this code ensures that `value`
and `origValue` will be identical (===). (The only reason they would
ever differ is that `value` is the result of calling `origValue`'s
`toJSON` method, if present. That invariant would throw if `toJSON`
were present.)

So, we can choose between them. Might as well use `origValue`. That
way, most of the code in this function uses `origValue`, which
clears the way for a nice upcoming refactor where we use a
switch/case instead of lots of `if` / `else if`s. It also helps
isolate `value` as something that's useful mainly for saving
computation in a particular way, as we do in the `Immutable.Map`
case (see bf8837c4a for more on that).