replaceRevive: Make an early return even earlier.
Not totally NFC because of the potential for a notable performance
gain (we now avoid doing some unnecessary checks).

This will allow us to refactor our many `if` / `else if` cases into
a simpler switch/case on `Object.getPrototypeOf(origValue)`, by
letting us call `Object.getPrototypeOf(origValue)` in the first
place. (It fails on `null` and `undefined`.)

This is NFC because `origValue` being non-`null` and having `typeof`
'object' is a necessary condition for any of the `if` / `else if`
conditions to be true.