typing status [nfc]: Factor out pmTypingKeyFromPmKeyIds.
There's no need to re-sort this list -- but that's true only because
the list is already sorted, in the same way we want to sort it for
this key.  So, add a helper function that takes advantage of that
fact, and put it within `recipient.js` which is the module where that
sort of fact is encapsulated.