UserAvatar [nfc]: Prepare to use React Hooks.
I'd like to try using react-redux's `useSelector` instead of
`connect` -- and that means turning this class component into a
function component. This is quite a simple component to convert.

I'm choosing a non-arrow function so that I can give the function a
name (the same name as the variable where we store it). I think
React (understandably) has trouble putting a name to arrow functions
and anonymous non-arrow functions in stack traces and in the dev
tools, though I haven't recently tested this.