redux-persist: Fix a recent regression in handling migrations.
08b77bb79f went a long way in reducing unnecessary work being done,
but unfortunately I wasn't careful enough about what happens with
migrations, and it introduced a source of hard-to-diagnose bugs that
might show up when migrations are added.

In particular, it's appropriate and necessary to turn around and
save the result of `REHYDRATE` back to the disk in cases where there
has been a migration. See discussion [1].

Rather than just revert 08b77bb79f, though, introduce a condition
that lets us get the benefit of the optimization in the far more
common case (a startup without a migration, vs. a startup with one).

The way we do that does break the separation between our
redux-persist and redux-persist-migrate, by making it so
redux-persist knows about the specifics of redux-persist-migrate.
But, as Greg says [2],

"""
I think that's fine, because a good "how to persistently store your
application's state" solution has to include migrations, and I think
the really high-quality libraries for it have ended up integrating a
specific way to do migrations. So breaking down the separation
between those two seems like the right architectural direction
anyway.
"""

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60dropCache.60/near/1093989
[2] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/.60dropCache.60/near/1094088