actions: Add ownUserId to FETCH_MESSAGES_COMPLETE.
We'll use this in a new reducer shortly.

There's a case where this can potentially change behavior: the thunk
actions that dispatch these actions do so after an `await`, which
means that by the time they get to calling `getOwnUserId` (after their
HTTP requests have returned) the user may have logged out, and it
might throw with "No server data found".

This is a case where we already have a worse bug, in that if the user
has switched to a *different* account and completed the initial fetch
for it by the time this request returns, then we'll be adding these
messages to our state to display alongside the messages that belong in
that account.  (Or try to, anyway: if the user IDs and stream IDs and
so on don't exist in the other account's realm, then we'll likely just
crash.)  We're aware of that bug and intend to fix it, but haven't
done the work for it yet.

See discussion:
  https://github.com/zulip/zulip-mobile/pull/4392#discussion_r554035895