PresenceStatusIndicator [nfc]: Get absolute-position style from props.
The three attribute settings (`bottom`, `right`, and `position`)
this removes from `opaqueBackground` aren't related to the
opaqueness of the background, so they seemed out-of-place. Instead,
I discovered, they seem to mimic the attributes set by one caller,
`UserAvatarWithPresence`, to position the status indicator in the
corner.

A consequence of setting those attributes in
`UserAvatarWithPresence` (and not allowing `props.style` to override
them) was that it changed `UserAvatarWithPresence`'s interface
slightly. Before, the caller could always control the positioning of
the status indicator (with `props.style`); after, it could not do so
if it was using the opaque background.

So, restore that bit of interface by taking the attributes from
`props.style`.

This way, we don't limit the options for our future selves or
others, who may want to add more callers that pass `true` for
`useOpaqueBackground` and specify their own positioning styles. This
commit is NFC because there aren't any such callsites now.