messagesReducer: Use `Immutable.Map#update` in one place.
Greg points out [1],

"""
A crucial feature here of `update` is that

> if the `updater` function returns the same value it was called
> with, then no change will occur.

So if the message isn't present, the updater will be passed
`undefined`, and just has to return `undefined` and the map will be
returned unchanged.
"""

See also, in that same comment, how Greg considers the effect on
clarity of relying on that subtle fact, and how that weighs with
other concerns, like efficiency.

[1] https://github.com/zulip/zulip-mobile/pull/4390#discussion_r555436082