contentHtmlFromPieceDescriptors [nfc]: Rename, from `renderMessagesAsHtml`.
The old name, with the vague word "render", lacked the specificity
needed to clearly say what part of the process this function
handles. The process of getting messages, headers, and time dividers
("content") to appear correctly on the screen has several steps, and
I wouldn't be surprised if readers had read `renderMessagesAsHtml`
as addressing a step it doesn't address, e.g., manipulating the DOM
by setting `innerHTML` on an element.

It was also misleading in that it assumed the input was only about
messages, when it's actually broader than that. Our "html piece
descriptors" also describe section headers and time dividers, not
just messages.

So, say more precisely what it does: take our special "html piece
descriptor" objects and return a string of HTML.